Expose the String substitution class.
